traffic: TrafficPhysicsInfo was 80 bytes too long -- the console stores the light-tag array as BYTES, and that one word shifted every member after it by 72
The brief asked whether traffic shares the six crash fixes the player's car just got. It does, and
that half is measured below. This commit is the one DEFECT the audit found on the traffic side, and
it is a layout fork of exactly the family the campaign has hit twice already (a struct forked at the
wrong SIZE beside the real record, so an array subscript walks the wrong stride).

=========================== 1. THE DEFECT, READ OFF THE ONLY WRITER ==========================
BrnTrafficEntityModule.h declared

    s32 maLightTagPointTypes[24];   // FLAG: retype to the real enum ...; the width does not change

The width DOES change. TrafficEntityModule::ProcessDeformationData @0x8271DEB0 is the only writer,
and its inner loop is a BYTE store with a stride of one:

    0x8271E92C  addi   r7, r7, -0x6E40   ; r7 = record + 0x60000 - 0x6E40 == record + 0x591C0
    0x8271E930  lvx128 v0, r0, r11       ; (the locator TRANSLATION row, stride 0x40)
    0x8271E93C  lwz    r6, 0(r9)         ; SOURCE stride 4 -- the 32-bit enum
    0x8271E944  addi   r9, r9, 4
    0x8271E948  stbx   r6, r7, r8        ; DESTINATION: stbx, a BYTE
    0x8271E94C  addi   r8, r8, 1         ; DESTINATION stride ONE

The 32-bit read is TRUNCATED into a byte lane. The array base is `this + 0x58210` == 360976
(RecordTrafficVehicleIsPhysical @0x82721144 `addis r3,r31,6` + @0x8272114C `addi r3,r3,-0x7DF0`),
so 0x591C0 - 0x58210 == 0xFB0 == 4016 is this array and 4016 + 24*1 == 4040 is the next member.
⚠️ The tree's own comment spelled that base 0x581D0 (== 360912, the decimal is right and the hex is
wrong); subtracting the wrong one puts the two direction floats at 0x101C, i.e. apparently PAST the
end of the record, and invents a bug that is not there. Corrected in place.

=========================== 2. WHAT IT COST, IN OFFSETS ======================================
24 entries x 3 surplus bytes = 72, plus align-16 tail padding = 80. Measured on the host by
compiling the struct, before and after:

    member                     was    now   console (and the store that pins it)
    maLightTagPointTypes      4016   4016   0xFB0   ProcessDeformationData stbx base 0x591C0
    mabWheelExists            4112   4040   0xFC8   BrnTrafficEntityModule_Render.cpp "info+4040+i"
    mfStuckTimeFront          4116   4044   0xFCC   UpdateVehicleStuckTimers @0x82708D90
    mfStuckTimeBack           4120   4048   0xFD0   UpdateVehicleStuckTimers @0x82708DA8
    mfSteeringDirection       4132   4060   0xFDC   RecordTrafficVehicleIsPhysical stfsx 0x591EC
    mfDrivingDirection        4136   4064   0xFE0   RecordTrafficVehicleIsPhysical stfsx 0x591F0
    miNumLightLocators        4140   4068   0xFE4   ProcessDeformationData        stbx  0x591F4
    muContactSideFlags        4176   4104   0x1008
    muOwningVehicleIndex      4178   4106   0x100A  TrafficPhysicsInfo::Construct @0x82751E88 sth
    sizeof                    4192   4112   0x1010  `mulli r10, r29, 0x1010` @0x82721130 AND
                                                    @0x8271E8E8, and 463776-360976 == 25*4112

Eleven static_asserts in BrnTrafficEntityModule.cpp now hold every one of those down. They are
ABSOLUTE offsets, which this file otherwise refuses on principle -- and they have to be, because the
console reaches this record by a HARD-CODED STRIDE rather than by a member name, and the record
contains no pointers, so the host layout can and must reproduce it exactly. All eleven pass under
tools\_gate_one.bat.

=========================== 3. WHY NOTHING CAUGHT IT ========================================
maTrafficPhysicsInfoList is a typed C array, so `&maTrafficPhysicsInfoList[i]` used the WRONG stride
CONSISTENTLY -- every write and every read agreed with each other and disagreed only with the
console. Nothing in the compile gate, the parity fingerprint or the faithfulness lint can see that.
It is not currently an out-of-bounds write; it is a correctness fork that becomes one the moment any
consumer uses the console's number. One already exists, dead: BrnReplays::TrafficPhysicsInfoRecord
(BrnReplayTrafficEntitySerialiserFrame.h:124) is `u8 maInfo[4112]` with
KU_PHYSICS_INFO_TOTAL == 25*4112, memcpy'd against the live array -- a 102,800-of-104,800 truncating
read with a per-record stride mismatch on top. It is unreachable today (UpdateSerialiser @0x8272DA80
has no body and BrnReplayTrafficEntitySerialiser.cpp is not mounted); after this commit the two
agree.

=========================== 4. WHAT THE SAME AUDIT MEASURED AND DID **NOT** CHANGE ===========
Recorded here because the next wave should not re-derive it. Runs trf_T1..trf_T5film, this exe:

* TRAFFIC IS ON EVERY ONE OF THE SIX FIXED PATHS. TrafficPhysics' concrete vtable is off_820D0C98
  (PhysicalTrafficManager::PhysicalTrafficManager @0x827E42E8 seats it last on all 20 bodies at a
  0x1430 stride), and it overrides NOTHING in slots +0x00..+0x2C -- every showtime/aftertouch
  predicate is the base body, read out of the image: +0x10/+0x14/+0x24 all == 0x827E2F38 `li r3,0`,
  +0x18 == 0x82C296C8 `li r3,1`, +0x1C == 0x827E24E8 (1.0f), +0x20 == 0x827EAB30 (0.0f),
  +0x28 == 0x8284CB38 `blr`. That is EXACTLY what VehiclePhysics.h:290-295 already declares, so
  there is no fifth wrong-slot defect here.
* MEASURED ON A TRAFFIC CAR, run trf_T3 (one deterministic 133 mph shot into stationary traffic):
  [tanbank] owner 2 fires 12 times at impRow 20.0 -- the CAR-CAR tangential row, which was dead on
  every car until 87dd1a61 -- and the player banks 12 of its own at 20.0 against it. Run trf_T4
  (95 s of ordinary driving) adds 19 more traffic applies at 20.0 and 45 at the 5.0 world row.
  [absorb] prints 29 crashing frames for that traffic car at absorption set 0.
* THE GATE IS THE CRASH STATE, AND IT IS THE CONSOLE'S. A traffic car only reaches mbCrashing
  through KU_RCTIR_CRASH_TRAFFIC, which DecideOutcomeOfRaceCarTrafficContact @0x825C70A0 raises only
  when the race car itself crashes on the impact. ShouldRaceCarCrashOnCarImpact @0x825C6FF8 is
  `impactSpeed * min(5000,otherMass)/victimMass * vulnerability > crashSpeedMPS * scale`, and the
  probe prints both sides: at 44 mph, stress 14.29 vs 38.00 -> SLAM (85 of 86 outcomes in trf_T4);
  at 133 mph, impactSpeed 57.72, stress 41.77 vs 38.00 -> CRASH, both cars. So "traffic does not
  crash hard" below ~82-117 mph closing is the console's own arithmetic, not a defect.
* AND ONCE IT DOES CRASH IT CRASHES HARD, measured: the struck car goes 0 -> 21.45 m/s with
  |w| = 6.14 rad/s in one frame while the player drops 74.5 -> 37.3 m/s ([T5-ram] f=6, maxMag
  20,562 N.s, sumMag 64,886 N.s); [tdef] reaches laneMax.z 0.500 m with 168 fatal rows and 98 moved
  verlet points; three panels come off, two of them HINGED; [wheel-probe] TWISTs one of its wheels.
  Frames scratch/flow_frames_trf5/bb_002950.bmp (intact) and bb_002954.bmp (the same hatchback four
  frames later, rear bumper torn in and the quarter creased) are the before/after.

Verified: both edited TUs pass tools\_gate_one.bat (exit 0), including the eleven new asserts.
Profile.sav hash 39950335... unchanged before and after every boot.
⚠️ Every run in this wave carried 131,869-164,186 copies of `IsInWorld()`
(GameSource/World/EntityModules/RaceCarEntityModule/BrnRaceCar.cpp:492), the rivals-wave assert
storm from be3749ba. Not touched, reported.

+++ src/GameSource/World/EntityModules/TrafficEntityModule/BrnTrafficEntityModule.h
@@ -319,10 +319,26 @@ namespace BrnTrafficIO { class InputBuffer_PreScene; class OutputBuffer_PreScene
         Matrix44Affine maWheelTransforms[KU_NUM_WHEELS];                  // :175
         Vector3 maLightLocatorPositions[KU_MAX_LIGHT_LOCATORS];           // :178
         // :179 -- BrnPhysics::Deformation::ETagPointType[24], stored as the enum's underlying
-        // 32-bit value so this header stays out of the deformation include graph.
-        // FLAG: retype to the real enum when a consumer needs it; the width does not change.
-        s32 maLightTagPointTypes[KU_MAX_LIGHT_LOCATORS];                  // :179
-        bool mabWheelExists[KU_NUM_WHEELS];                               // :181
+        // value so this header stays out of the deformation include graph.
+        //
+        // ⭐⭐⭐ WIDTH CORRECTED 2026-09-06 (traffic wave). This was `s32[24]` under a FLAG that
+        // said "the width does not change". It DOES: the console stores ONE BYTE per entry, and
+        // that made the whole record 80 bytes too long on the host (4192 vs the console's 4112),
+        // shifting EVERY member after it by 72. ProcessDeformationData @0x8271DEB0, the only
+        // writer, read instruction by instruction:
+        //     0x8271E92C  addi   r7, r7, -0x6E40   ; r7 = record + 0x60000 - 0x6E40 = +0x591C0
+        //     0x8271E93C  lwz    r6, 0(r9)         ; SOURCE stride 4 (the 32-bit enum)
+        //     0x8271E944  addi   r9, r9, 4
+        //     0x8271E948  stbx   r6, r7, r8        ; DESTINATION: a BYTE
+        //     0x8271E94C  addi   r8, r8, 1         ; DESTINATION stride ONE
+        // The array base is `this + 0x58210` (== 360976; 0x82721144 `addis r3,r31,6` +
+        // 0x8272114C `addi r3,r3,-0x7DF0`), so 0x591C0 - 0x58210 == 0xFB0 == 4016 is this array
+        // and 4016 + 24*1 == 4040 is mabWheelExists -- which BrnTrafficEntityModule_Render.cpp
+        // already documents at "info+4040+i", and which puts mfStuckTimeFront back on the
+        // +0xFCC (4044) that UpdateVehicleStuckTimers @0x82708D90 pins. With s32 it sat at 4116.
+        // The static_asserts in BrnTrafficEntityModule.cpp hold all of this down now.
+        u8 maLightTagPointTypes[KU_MAX_LIGHT_LOCATORS];                   // :179 (X360 BYTE)
+        bool mabWheelExists[KU_NUM_WHEELS];                               // :181 (X360 +0xFC8)
         f32 mfStuckTimeFront;                                             // :183
         f32 mfStuckTimeBack;                                              // :184
         f32 mfStuckTimerDebounce;                                         // :185
